+++ R/utils-mspeaklists.R
@@ -253,6 +253,29 @@ assignPrecursorToMSPeakList <- function(MSPeakList, precursorMZ)
     return(MSPeakList)
 }
 
+assignMSPLIDs <- function(MSPeakLists)
+{
+    addIDs <- function(pl)
+    {
+        add <- function(p)
+        {
+            p <- copy(p)
+            p[, ID := seq_len(.N)]
+            setcolorder(p, "ID")
+            return(p)
+        }
+        if (!is.null(pl[["MS"]]))
+            pl$MS <- add(pl$MS)
+        if (!is.null(pl[["MSMS"]]))
+            pl$MSMS <- add(pl$MSMS)
+        
+        return(pl)
+    }
+    MSPeakLists@peakLists <- lapply(MSPeakLists@peakLists, function(pla) lapply(pla, addIDs))
+    MSPeakLists@averagedPeakLists <- lapply(MSPeakLists@averagedPeakLists, addIDs)
+    return(MSPeakLists)
+}
+
 deIsotopeMSPeakList <- function(MSPeakList, negate)
 {
     # UNDONE: this is DA specific and not tested for a while, maybe update some day...